// source/types/constants/PromoObjective.ts
export enum PromoObjective {
    PROMO_OBJECTIVE_VIEWS_PHOTO = "PROMO_OBJECTIVE_VIEWS_PHOTO",
    PROMO_OBJECTIVE_VIEWS_VIDEO = "PROMO_OBJECTIVE_VIEWS_VIDEO",
    PROMO_OBJECTIVE_TRAFFIC_WEBSITE = "PROMO_OBJECTIVE_TRAFFIC_WEBSITE",
    PROMO_OBJECTIVE_TRAFFIC_SOCIAL_MEDIA = "PROMO_OBJECTIVE_TRAFFIC_SOCIAL_MEDIA",
    PROMO_OBJECTIVE_ACTION_LIKE = "PROMO_OBJECTIVE_ACTION_LIKE",
    PROMO_OBJECTIVE_ACTION_SUBSCRIPTION = "PROMO_OBJECTIVE_ACTION_SUBSCRIPTION",
    PROMO_OBJECTIVE_ACTION_DOWNLOAD = "PROMO_OBJECTIVE_ACTION_DOWNLOAD",
    PROMO_OBJECTIVE_ACTION_APP_INSTALL = "PROMO_OBJECTIVE_ACTION_APP_INSTALL",
    PROMO_OBJECTIVE_PURCHASE_REGISTRATION = "PROMO_OBJECTIVE_PURCHASE_REGISTRATION",
    PROMO_OBJECTIVE_PURCHASE_SALE = "PROMO_OBJECTIVE_PURCHASE_SALE",
}